Extract schema procedure computation to dedicated class
