+++ spark-cypher/src/main/scala/org/opencypher/spark/api/io/neo4j/Neo4jReadOnlySource.scala
@@ -28,22 +28,17 @@ package org.opencypher.spark.api.io.neo4j
 
 import org.apache.spark.sql.DataFrame
 import org.apache.spark.sql.types.StructType
-import org.neo4j.driver.v1.Value
-import org.neo4j.driver.v1.util.Function
 import org.opencypher.okapi.api.graph.GraphName
 import org.opencypher.okapi.api.schema.Schema
-import org.opencypher.okapi.api.types.CypherType
+import org.opencypher.okapi.impl.exception.UnsupportedOperationException
 import org.opencypher.okapi.impl.schema.SchemaImpl
 import org.opencypher.spark.api.CAPSSession
 import org.opencypher.spark.api.io.AbstractDataSource
 import org.opencypher.spark.api.io.metadata.CAPSGraphMetaData
-import org.opencypher.spark.api.io.neo4j.Neo4jReadOnlySource.{metaPrefix, schemaProcedureName}
-import org.opencypher.spark.impl.io.neo4j.Neo4jHelpers._
+import org.opencypher.spark.api.io.neo4j.Neo4jReadOnlySource.metaPrefix
 import org.opencypher.spark.impl.io.neo4j.external.Neo4j
 import org.opencypher.spark.schema.CAPSSchema
-
-import scala.collection.JavaConverters._
-import scala.util.{Failure, Success, Try}
+import org.opencypher.spark.schema.CAPSSchema._
 
 
 /**
@@ -61,67 +56,6 @@ case class Neo4jReadOnlySource(
 )(implicit session: CAPSSession)
   extends AbstractDataSource {
 
-  private def computeSchema(name: GraphName): Option[Schema] = {
-    Try {
-      config.cypher("CALL dbms.procedures() YIELD name AS name").exists { map =>
-        map("name").value == schemaProcedureName
-      }
-    } match {
-      case Success(true) =>
-        schemaFromProcedure(config)
-      case Success(false) =>
-        System.err.println("Neo4j schema procedure not activated. Consider activating the procedure `" + schemaProcedureName + "`.")
-        None
-      case Failure(error) =>
-        System.err.println(s"Retrieving the procedure list from the Neo4j database failed: $error")
-        None
-    }
-  }
-
-  private[neo4j] def schemaFromProcedure(config: Neo4jConfig): Option[Schema] = {
-    Try {
-      val rows = config.execute { session =>
-        val result = session.run("CALL " + schemaProcedureName)
-
-        result.list().asScala.map { row =>
-          val typ = row.get("type").asString()
-          val labels = row.get("nodeLabelsOrRelType").asList(new Function[Value, String] {
-            override def apply(v1: Value): String = v1.asString()
-          }).asScala.toList
-
-          row.get("property").asString() match {
-            case "" => // this label/type has no properties
-              (typ, labels, None, None)
-            case property =>
-              val typeString = row.get("cypherType").asString()
-              val cypherType = CypherType.fromName(typeString)
-              (typ, labels, Some(property), cypherType)
-          }
-        }
-      }
-
-      rows.groupBy(row => row._1 -> row._2).map {
-        case ((typ, labels), tuples) =>
-          val properties = tuples.collect {
-            case (_, _, p, t) if p.nonEmpty => p.get -> t.get
-          }
-
-          typ match {
-            case "Node" =>
-              Schema.empty.withNodePropertyKeys(labels: _*)(properties: _*)
-            case "Relationship" =>
-              Schema.empty.withRelationshipPropertyKeys(labels.headOption.getOrElse(""))(properties: _*)
-          }
-      }.foldLeft(Schema.empty)(_ ++ _)
-    } match {
-      case Success(schema) => Some(schema)
-      case Failure(error) =>
-        System.err.println(s"Could not load schema from Neo4j: ${error.getMessage}")
-        error.printStackTrace()
-        None
-    }
-  }
-
   override def tableStorageFormat: String = "neo4j"
 
   override protected def listGraphNames: List[String] = List(entireGraph.value) // TODO: extend with meta label approach
@@ -132,7 +66,10 @@ case class Neo4jReadOnlySource(
   }
 
   override protected def readSchema(graphName: GraphName): CAPSSchema = {
-    val graphSchema = computeSchema(graphName).get
+    val graphSchema = SchemaFromProcedure(config) match {
+      case None => throw UnsupportedOperationException("needs the procedure")
+      case Some(schema) => schema
+    }
     val filteredSchema = getMetaLabel(graphName) match {
       case None =>
         graphSchema
@@ -141,8 +78,6 @@ case class Neo4jReadOnlySource(
         val labelPropertyMap = graphSchema.labelPropertyMap.filterForLabels(metaLabel)
         SchemaImpl(labelPropertyMap, graphSchema.relTypePropertyMap)
     }
-
-    import org.opencypher.spark.schema.CAPSSchema._
 
     filteredSchema.asCaps
   }
@@ -220,5 +155,4 @@ case class Neo4jReadOnlySource(
 
 object Neo4jReadOnlySource {
   val metaPrefix: String = "___"
-  val schemaProcedureName = "org.neo4j.morpheus.procedures.schema"
 }
